Builtin/non-builtin types on method invocation and `New-Object`
Before, type literals only worked w/ builtins. Now you can do:

    > [System.String]::Concat( 'ab', '12' )

Before, `New-Object` didn't recognize builtin types. Now you can do:

    > New-Object bool
